fixed syntax error
unsubMessages is no longer visible to useEffect() clean-up function: the variable’s scope is now limited to the if (isConnected) {...} block.

Need to declare the variable outside the if statement and then assign it w/ the onSnapshot() listener inside the if statement block, as before. Since unsubMessages is now declared and initialised on separate lines, you’ll have to use let instead of const